requirements.txt
streamlit
chromadb
google-generativeai